api/requirements.txt
cchardet
pylint
httpie
rednose
codecov
cerberus
click
cffi
Flask
Flask-API
flasgger
itsdangerous==0.24
Jinja2
kombu
MarkupSafe
pytz
pytest-flask
vine
werkzeug
gunicorn==20.0.4
requests
six==1.9.0
